aries/requirements.txt
# Core Django
Django==5.1.4
djangorestframework==3.16.1
django-filter==25.1
django-cors-headers==4.9.0
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# Email
resend==0.8.0

# Image processing
Pillow==10.1.0

# Country fields
django-countries==7.6.1

# Testing
pytest==7.4.3
pytest-django==4.7.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Security
python-decouple==3.8